broken but the direction should be to finish the update method for the manager and all its sideeffects.
GETTO: Implemetn updateproject in manager and tests.
