+++ Taskter/Utilities/Domain/Request/Project/ProjectUpdateRequest.cs
@@ -5,12 +5,12 @@ namespace Utilities.Taskter.Domain
     public class ProjectUpdateRequest
     {
         /// <summary>
-        /// The name of the project.
+        /// The name of the project. Could be changed if desired.
         /// </summary>
         public string Name { get; set; } = string.Empty;
 
         /// <summary>
-        /// The project acronym, used as an identifier.
+        /// The project acronym, used as new identifier.
         /// </summary>
         public string ProjectAcronym { get; set; } = string.Empty;
 
@@ -18,5 +18,6 @@ namespace Utilities.Taskter.Domain
         /// The existing project acronym, used as an identifier.
         /// </summary>
         public string ExistingProjectAcronym { get; set; } = string.Empty;
+
     }
 }
